requirements.txt
streamlit>=1.28.0
llama-index>=0.9.0
llama-index-readers-google>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pandas>=1.5.0
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0